# GameWizOpenApiSpec.yaml
openapi: 3.0.4
info:
  title: GameWizApi
  description: |-
    This is an OpenApi 3.0 specification for Game Wiz api. A server used for asynchronous games. GWA is a small, solo,
    project by mfv.
  
  contact:
    email: mfvdev@icloud.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/v0
tags:
  - name: health
    description: endpoint for checking health of api
  - name: user
    description: Operations for users
  - name: session
    description: operations for game sessions
paths:
  /health:
    get:
      tags:
        - health
      summary: Useable for devs only, pings server for health check.
      description: Server health status
      operationId: health
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'
        '400':
          description: Invalid Request
        '403':
          description: Forbidden
        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user:
    post:
      tags:
        - user
      summary: Create a new user
      description: Endpoint for registering new user
      operationId: createAccount
      requestBody:
        description: Create new user account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
        required: true
      responses:
        '201':
          description: Create Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid Request
        '403':
          description: Forbidden
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /session/{gameId}:
    get:
      parameters:
        - in: path
          name: gameId
          schema:
            type: string
          required: true
      tags:
        - session
      summary: get session data
      description: endpoint for getting next player move, game state
      operationId: getSession
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid Request
        '403':
          description: Forbidden
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /session:
    post:
      tags:
        - session
      summary: Create game session
      description: operation for new session
      operationId: CreateSession
      requestBody:
        description: Create new game session
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSession'
        required: true
      responses:
        '201':
          description: Create Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid Request
        '403':
          description: Forbidden
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    HealthCheck:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    CreateSession:
      type: object
      required:
        - gameId
        - maxPlayerCount
        - minPlayerCount
        - hostUID
        - upNext
      properties:
        gameId:
          type: string
        maxPlayerCount:
          type: integer
          format: int32
        minPlayerCount:
          type: integer
          format: int32
        hostUID:
          type: string
        upNext:
          type: string
        metaData:
          type: object
          additionalProperties: {}
    Session:
      type: object
      required:
        - sessionId
        - gameId
        - maxPlayerCount
        - minPlayerCount
        - players
        - upNext
        - status
        - gameState
      properties:
        sessionId:
          type: string
          example: "apseion64565dsaf"
        gameId:
          type: string
          example: "TicTacToe"
        maxPlayerCount:
          type: integer
        minPlayerCount:
          type: integer
        players:
          type: array
          items:
            type: string
          example: ["player69"]
        upNext:
          type: string
          example: "player69"
        metaData:
          type: object
          additionalProperties: {}
        gameState:
          type: object
          additionalProperties: {}
        status:
          $ref: '#/components/schemas/Status'
    CreateAccount:
      type: object
      required:
        - username
        - password
        - email
        - birthdate
      properties:
        username:
          type: string
          example: letmesolohim
        password:
          type: string
          pattern: "^[A-Za-z\\d@$!%*?&_+]{8,}$" #TODO, check regex
          example: 'somePa$$w0rd_'
          format: password
        email:
          type: string
          pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+[a-z]{3}$"
          example: 'someEmail@provider.com'
          format: email
        birthdate:
          type: integer
          format: int64
          example: 6584613516
      xml:
        name: playercreate
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    Status:
      type: string
      enum:
        - WAITING_FOR_PLAYERS
        - IN_PROGRESS
        - TERMINATED
        - FINISHED
  requestBodies:
    CreateAccount:
      description: User object to add to user store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateAccount'
        application/xml:
          schema:
            $ref: '#/components/schemas/CreateAccount'
    CreateSession:
      description: Request to create new game session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateSession'
        application/xml:
          schema:
            $ref: '#/components/schemas/CreateSession'